Reorganized mysql to Vertica import code.
