+++ edx/analytics/tasks/tests/acceptance/test_internal_reporting_database.py
