+++ setup.cfg
@@ -69,6 +69,7 @@ edx.analytics.tasks =
     obfuscation = edx.analytics.tasks.obfuscation:ObfuscatedCourseTask
     push_to_vertica_lms_courseware_link_clicked = edx.analytics.tasks.lms_courseware_link_clicked:PushToVerticaLMSCoursewareLinkClickedTask
     load-course-catalog = edx.analytics.tasks.load_internal_reporting_course_catalog:PullCourseCatalogAPIData
+    load-internal-database = edx.analytics.tasks.load_internal_reporting_database:ImportMysqlToVerticaTask
 
 
 mapreduce.engine =
